:pencil: delete confirmation dialog

+++ src/components/TableItem.tsx
@@ -7,6 +7,8 @@ import {
 } from "../helpers/firebaseFunctions";
 import { collection, getFirestore, query, where } from "firebase/firestore";
 import { app } from "../db/Firebase";
+import { HandleRadioCheckValueFunction } from "../App";
+import { ConfirmDialog } from "primereact/confirmdialog";
 
 type Props = {
   item?: Item;
@@ -16,6 +18,10 @@ type Props = {
   windowSize: windowSizeProps;
   pathFirebase: PathFirebaseProps;
   userId: string;
+  handleRadioChecked: ({
+    redditoValue,
+    spesaValue,
+  }: HandleRadioCheckValueFunction) => void;
 };
 
 // INITIALIZING FIRESTORE DB
@@ -29,11 +35,13 @@ export function TableItem({
   windowSize,
   pathFirebase,
   userId,
+  handleRadioChecked,
 }: Props) {
+  const [visible, setVisible] = useState(false);
   if (isSettings) {
     null;
   } else if (isCategorySettings) {
-    const bgColor = category?.expense ? "bg-red-500" : "bg-green-500";
+    const bgColor = category?.expense ? "bg-red-700" : "bg-green-700";
     return (
       <tr className="flex w-full px-5 p-3 items-center">
         <td className="flex w-2/3 py-3 menuBreak:px-2 text-left">
@@ -47,11 +55,24 @@ export function TableItem({
           </div>
         </td>
         <td className="flex w-1/6 py-3 menuBreak:px-2 justify-end">
+          <ConfirmDialog
+            className="bg-red-800 p-5 border border-red-800 rounded-3xl flex gap-5 items-center w-64"
+            visible={visible}
+            onHide={() => setVisible(false)}
+            message={`Vuoi eliminare ${category!.title}?`}
+            header="Conferma"
+            icon={""}
+            accept={() => {
+              deleteFirebaseData(userId, "categories", category!.id),
+                handleRadioChecked({ redditoValue: false, spesaValue: false });
+            }}
+            reject={() => setVisible(false)}
+          />
           <div
             className="w-10 text-right py-1 text-lg cursor-pointer"
-            onClick={() =>
-              deleteFirebaseData(userId, "categories", category!.id)
-            }
+            onClick={() => {
+              setVisible(true);
+            }}
           >
             ⛔
           </div>
@@ -87,8 +108,8 @@ export function TableItem({
         });
       }
     }, [item]);
-    const bgColor = itemCategory?.expense ? "bg-red-500" : "bg-green-500";
-    const textColor = itemCategory?.expense ? "text-red-500" : "text-green-500";
+    const bgColor = itemCategory?.expense ? "bg-red-700" : "bg-green-700";
+    const textColor = itemCategory?.expense ? "text-red-600" : "text-green-600";
     return (
       <tr className="flex px-5 justify-between items-center">
         <td className="flex w-2/12 py-3 px-0 menuBreak:px-2 text-left">
@@ -115,9 +136,24 @@ export function TableItem({
           </div>
         </td>
         <td className="flex w-1/12 py-3 px-0 menuBreak:px-2 text-left ">
+          <ConfirmDialog
+            className="bg-red-800 p-5 border border-red-800 rounded-3xl flex gap-5 items-center w-64"
+            visible={visible}
+            onHide={() => setVisible(false)}
+            message={`Vuoi eliminare ${item!.title}?`}
+            header="Conferma"
+            icon={""}
+            accept={() => {
+              deleteFirebaseData(userId, "items", item!.id),
+                handleRadioChecked({ redditoValue: false, spesaValue: false });
+            }}
+            reject={() => setVisible(false)}
+          />
           <div
-            className="w-full text-right py-1 text-lg cursor-pointer"
-            onClick={() => deleteFirebaseData(userId, "items", item!.id)}
+            className="w-10 text-right py-1 text-lg cursor-pointer"
+            onClick={() => {
+              setVisible(true);
+            }}
           >
             ⛔
           </div>
